Refactor
1. Move sendTelemetryData function to common.
2. Update named imports in common.
