+++ docs-preview/src/helper/common.ts
@@ -1,13 +1,14 @@
 "use-strict";
 
-import * as vscode from "vscode";
+import { Uri, window, workspace } from "vscode";
+import { reporter } from "./telemetry";
 
 /**
  * Create a posted warning message and applies the message to the log
  * @param {string} message - the message to post to the editor as an warning.
  */
 export function postWarning(message: string) {
-    vscode.window.showWarningMessage(message);
+    window.showWarningMessage(message);
 }
 
 /**
@@ -15,7 +16,7 @@ export function postWarning(message: string) {
  * @param {string} message - the message to post to the editor as an information.
  */
 export function postInformation(message: string) {
-    vscode.window.showInformationMessage(message);
+    window.showInformationMessage(message);
 }
 
 /**
@@ -23,14 +24,14 @@ export function postInformation(message: string) {
  * @param {string} message - the message to post to the editor as an information.
  */
 export function postError(message: string) {
-    vscode.window.showErrorMessage(message);
+    window.showErrorMessage(message);
 }
 
 export function hasValidWorkSpaceRootPath(senderName: string) {
-    const folderPath = vscode.workspace.rootPath;
+    const folderPath = workspace.rootPath;
 
     if (folderPath == null) {
-        postWarning("The " + senderName + " command requires an active workspace. Please open VS Code from the root of your clone to continue.");
+        postWarning(`The ${senderName} command requires an active workspace. Please open VS Code from the root of your clone to continue.`);
         return false;
     }
 
@@ -52,11 +53,19 @@ export function generateTimestamp() {
  * Return repo name
  * @param Uri
  */
-export function getRepoName(workspacePath: vscode.Uri) {
+export function getRepoName(workspacePath: Uri) {
     // let repoName;
-    const repo = vscode.workspace.getWorkspaceFolder(workspacePath);
+    const repo = workspace.getWorkspaceFolder(workspacePath);
     if (repo) {
         const repoName = repo.name;
         return repoName;
     }
+}
+
+export function sendTelemetryData(telemetryCommand: string, commandOption: string) {
+    const editor = window.activeTextEditor;
+    const workspaceUri = editor.document.uri;
+    const activeRepo = getRepoName(workspaceUri);
+    const telemetryProperties = activeRepo ? { command_option: commandOption, repo_name: activeRepo } : { command_option: commandOption, repo_name: "" };
+    reporter.sendTelemetryEvent(telemetryCommand, telemetryProperties);
 }
